Add files via upload
In this project worked on utilizing NLP (Natural Language Processing) and the Pickle package for detecting phishing websites. Phishing attacks continue to be a significant threat to individuals and organizations worldwide, and detecting and preventing them is crucial. Through this project, I was able to develop a machine learning model that uses NLP techniques to analyze the language used on websites and identify potential phishing attempts. The model was trained on a large dataset of known phishing websites and achieved high accuracy in detecting such sites. This approach has enormous potential to benefit cybersecurity and protect individuals and businesses from phishing attacks. I am thrilled to have had the opportunity to work on this project and contribute to the field of cybersecurity. I look forward to exploring more applications of NLP and machine learning in cybersecurity and continuing to advance the capabilities of artificial intelligence in detecting and preventing cyber threats.